【维护】代码维护，Scene创建实体方法修改，Entity添加组件方法修改。camke指定UTF8（貌似还有警告

+++ Engine/src/Himii/Renderer/Buffer.h
@@ -130,8 +130,8 @@ namespace Himii
             for (auto &element: m_Elements)
             {
                 element.Offset = offset;
-                offset += element.Size;
-                m_Stride += element.Size;
+                offset += static_cast<uint32_t>(element.Size);
+                m_Stride += static_cast<uint32_t>(element.Size);
             }
         }
